Added all repositories with bare minimum crud operations

+++ DTO/TradingDeal.cs
@@ -11,6 +11,6 @@ namespace DTO
         public string Id { get; set; }
         public string CardToTrade { get; set; }
         public string Type { get; set; }
-        public float MinimumDamage { get; set; }
+        public int MinimumDamage { get; set; }
     }
 }
